fixing sonarcube complaoints
