+++ src/app-components/TimePicker/TimeSegment/TimeSegment.tsx
@@ -5,26 +5,7 @@ import { Textfield } from '@digdir/designsystemet-react';
 import { useSegmentDisplay } from 'src/app-components/TimePicker/TimeSegment/hooks/useSegmentDisplay';
 import { useSegmentInputHandlers } from 'src/app-components/TimePicker/TimeSegment/hooks/useSegmentInputHandlers';
 import { useTypingBuffer } from 'src/app-components/TimePicker/TimeSegment/hooks/useTypingBuffer';
-import type { TimeFormat } from 'src/app-components/TimePicker/TimePicker';
-import type { SegmentType } from 'src/app-components/TimePicker/utils/keyboardNavigation';
-
-export interface TimeSegmentProps {
-  value: number | string;
-  min: number;
-  max: number;
-  type: SegmentType;
-  format: TimeFormat;
-  onValueChange: (value: number | string) => void;
-  onNavigate: (direction: 'left' | 'right') => void;
-  placeholder?: string;
-  disabled?: boolean;
-  readOnly?: boolean;
-  required?: boolean;
-  'aria-label': string;
-  'aria-describedby'?: string;
-  className?: string;
-  autoFocus?: boolean;
-}
+import type { TimeSegmentProps } from 'src/app-components/TimePicker/types';
 
 export const TimeSegment = React.forwardRef<HTMLInputElement, TimeSegmentProps>(
   (
